requirements.txt
langchain
langchain-openai
openpyxl
pandas
jinja2
python-dotenv
